refactor(cli/migrate): rename input utility files to `input-<utility>.ts`
