+++ packages/remix-dev/cli/migrate/run.ts
@@ -2,7 +2,7 @@ import { checkGitStatus } from "../check-git-status";
 import type { migrationOptions } from "./migration-options";
 import type { Transform, TransformArgs } from "./transforms";
 import { updateRemixImports } from "./transforms";
-import { checkProjectDir, checkMigration } from "./check";
+import { checkProjectDir, checkMigration } from "./input-check";
 
 const transformFunctionByName: Record<
   typeof migrationOptions[number]["value"],
